requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
openai
python-dotenv
yfinance
pandas
openpyxl
streamlit
